Refactor components file to their modules folder
