+++ src/modules/home/ui/components/search-filters/categories-sidebar.tsx
