+++ src/modules/home/ui/components/search-filters/categories.tsx
